Allow custom Checker implementations
