+++ src/Exceptions/SignerException.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+namespace BitWasp\Bitcoin\Exceptions;
+
+class SignerException extends \Exception
+{
+
+}
